added a subproject BatteryTool - command line interface for calculation tortuosity

+++ code/batteryviz/src/BatteryApp.h
@@ -80,7 +80,7 @@ protected:
 	> _saEllipsoid;
 
 
-	blib::DiffusionSolver<float> _diffSolver;
+	blib::DiffusionSolver<double> _diffSolver;
 
 
 	friend Ui;
